refactor: fully decouple plot_examples and move create_dataarray to BaseDatastore
